Dakota_Jameson_AutomatedSweep_Improved.m
This program is broken up into different parts. First, the program executes the automated voltage sweep and records the measured voltage and current that the DMM is reading. Next, the program creates an I-V graph with a line of best fit to establish the relationship between voltage and current and verify Ohm's law based on that relationship. Third, the program checks to see if the measured max resistance falls within the range of an actual 1k ohm resistor.